+++ test/captions.spec.js
@@ -1,15 +1,15 @@
-const apiVideo = require('../lib');
 const path = require('path');
 const expect = require('chai').expect;
+const apiVideo = require('../lib');
 
 describe('Captions ressource', () => {
   describe('Upload', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
-      let source = path.join(__dirname, 'data/en.vtt');
-      let properties = {
+      const client = new apiVideo.Client({ apiKey: 'test' });
+      const source = path.join(__dirname, 'data/en.vtt');
+      const properties = {
         videoId: 'vix1x1x1x1x1x1x1x1x1x',
-        language: 'en'
+        language: 'en',
       };
       client.captions.upload(source, properties);
       expect(client.captions.browser.lastRequest).to.deep.property('url', 'https://ws.api.video/videos/vix1x1x1x1x1x1x1x1x1x/captions/en');
@@ -21,69 +21,69 @@ describe('Captions ressource', () => {
 
   describe('updateDefault', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
-      let videoId = 'vix1x1x1x1x1x1x1x1x1x';
-      let language = 'en';
+      const client = new apiVideo.Client({ apiKey: 'test' });
+      const videoId = 'vix1x1x1x1x1x1x1x1x1x';
+      const language = 'en';
       client.captions.updateDefault(videoId, language, true);
       expect(client.captions.browser.lastRequest).to.deep.equal({
         url: 'https://ws.api.video/videos/vix1x1x1x1x1x1x1x1x1x/captions/en',
         method: 'PATCH',
         headers: {},
         body: { default: true },
-        json: true
+        json: true,
       });
     });
   });
 
   describe('get', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
+      const client = new apiVideo.Client({ apiKey: 'test' });
       client.captions.get('vix1x1x1x1x1x1x1x1x1x', 'en');
       expect(client.captions.browser.lastRequest).to.deep.equal({
         url: 'https://ws.api.video/videos/vix1x1x1x1x1x1x1x1x1x/captions/en',
         method: 'GET',
         headers: {},
-        json: true
+        json: true,
       });
     });
   });
 
   describe('getAll', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
+      const client = new apiVideo.Client({ apiKey: 'test' });
       client.captions.getAll('vix1x1x1x1x1x1x1x1x1x');
       expect(client.captions.browser.lastRequest).to.deep.equal({
         url: 'https://ws.api.video/videos/vix1x1x1x1x1x1x1x1x1x/captions',
         method: 'GET',
         headers: {},
-        json: true
+        json: true,
       });
     });
   });
 
   describe('delete', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
+      const client = new apiVideo.Client({ apiKey: 'test' });
       client.captions.delete('vix1x1x1x1x1x1x1x1x1x', 'en');
       expect(client.captions.browser.lastRequest).to.deep.equal({
         url: 'https://ws.api.video/videos/vix1x1x1x1x1x1x1x1x1x/captions/en',
         method: 'DELETE',
         headers: {},
-        json: true
+        json: true,
       });
     });
   });
 
   describe('cast', () => {
     it('Should return caption object', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
-      let data = {
+      const client = new apiVideo.Client({ apiKey: 'test' });
+      const data = {
         uri: '/videos/vix1x1x1x1x1x1x1x1x1x/captions/en',
         src: 'https://cdn.api.video/vod/vix1x1x1x1x1x1x1x1x1x/token/1390f745-a267-493b-8a2d-a8ea0e017042/captions/en.vtt',
-        srclang: "en",
-        default: true
+        srclang: 'en',
+        default: true,
       };
-      let caption = client.captions.cast(data);
+      const caption = client.captions.cast(data);
       expect(caption).to.deep.equal(data);
     });
   });