+++ test/tokens.spec.js
@@ -1,10 +1,10 @@
+const { expect } = require('chai');
 const apiVideo = require('../lib');
-const expect = require('chai').expect;
 
 describe('Tokens ressource', () => {
   describe('generate', () => {
     it('Sends good request', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
+      const client = new apiVideo.Client({ apiKey: 'test' });
       client.tokens.generate().catch((error) => {
         console.log(error);
       });
@@ -13,18 +13,17 @@ describe('Tokens ressource', () => {
         method: 'POST',
         headers: {},
         body: {},
-        json: true
+        json: true,
       });
     });
   });
 
   describe('cast', () => {
     it('Should return caption object', () => {
-      const client = new apiVideo.Client({username: 'test', apiKey: 'test'});
-      let data = {token: 'tox1x1x1x1x1x1x1x1x1x'};
-      let token = client.tokens.cast(data);
+      const client = new apiVideo.Client({ apiKey: 'test' });
+      const data = { token: 'tox1x1x1x1x1x1x1x1x1x' };
+      const token = client.tokens.cast(data);
       expect(token).to.deep.equal(data.token);
     });
   });
-
 });